+++ ogg_player.h
@@ -22,8 +22,8 @@ struct OggPlayer {
 	void pauseTrack();
 	void resumeTrack();
 	bool isPlaying() const { return _impl != 0; }
-	bool mix(int8_t *buf, int len);
-	static bool mixCallback(void *param, int8_t *buf, int len);
+	bool mix(int16_t *buf, int len);
+	static bool mixCallback(void *param, int16_t *buf, int len);
 
 	Mixer *_mix;
 	FileSystem *_fs;
