+++ systemstub.h
@@ -45,7 +45,7 @@ struct PlayerInput {
 };
 
 struct SystemStub {
-	typedef void (*AudioCallback)(void *param, int8_t *stream, int len);
+	typedef void (*AudioCallback)(void *param, int16_t *stream, int len);
 
 	PlayerInput _pi;
